third_party/blink/web_tests/http/tests/preload/warning/blob-request-expected.txt

CONSOLE WARNING: A preload for 'http://127.0.0.1:8000/preload/warning/resources/hello.txt' is found, but is not used because the new request loads the content as a blob.
This is a testharness.js-based test.
Harness: the test ran to completion.
